+++ src/platforms/arm/mgm240/fastled_arm_mgm240.h
@@ -3,3 +3,7 @@
 #include "platforms/arm/mgm240/led_sysdefs_arm_mgm240.h"
 #include "platforms/arm/mgm240/fastpin_arm_mgm240.h"
 #include "platforms/arm/mgm240/clockless_arm_mgm240.h"
+
+// Include ezWS2812 controllers
+#include "platforms/arm/mgm240/clockless_ezws2812_gpio.h"
+#include "platforms/arm/mgm240/clockless_ezws2812_spi.h"